# Improve performance/reduce allocations in ButtonInput