Follows on from the previous performance improvements, which now allows `ButtonInput` to use the builder and forego an extra `.ToArray()` + allocations therein.

Master:

|        Method | count |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |------ |-----------:|---------:|---------:|-------:|------:|------:|----------:|
|         Apply |     ? | 1,060.5 ns |  4.93 ns |  4.61 ns | 0.0706 |     - |     - |    2416 B |
| FromTwoStates |     1 |   370.1 ns |  2.30 ns |  2.04 ns | 0.0219 |     - |     - |     744 B |
| FromTwoStates |     5 |   849.8 ns |  4.34 ns |  3.85 ns | 0.0448 |     - |     - |    1528 B |
| FromTwoStates |    10 | 1,439.6 ns |  5.94 ns |  5.27 ns | 0.0820 |     - |     - |    2752 B |
| FromTwoStates |    50 | 5,172.2 ns | 19.83 ns | 18.55 ns | 0.3052 |     - |     - |   10416 B |

This PR:

|        Method | count |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |------ |-----------:|---------:|---------:|-------:|------:|------:|----------:|
|         Apply |     ? | 1,050.6 ns | 11.43 ns | 10.13 ns | 0.0706 |     - |     - |    2416 B |
| FromTwoStates |     1 |   298.9 ns |  2.10 ns |  1.97 ns | 0.0205 |     - |     - |     688 B |
| FromTwoStates |     5 |   724.4 ns |  3.61 ns |  3.38 ns | 0.0362 |     - |     - |    1232 B |
| FromTwoStates |    10 | 1,261.8 ns |  6.70 ns |  6.27 ns | 0.0648 |     - |     - |    2176 B |
| FromTwoStates |    50 | 4,906.3 ns | 18.98 ns | 16.83 ns | 0.2441 |     - |     - |    8256 B |